Refactor tasks to use shared DB dependency
