+++ backend/src/askpolis/data_fetcher/tasks.py
@@ -1,20 +1,15 @@
-import os
-
 from celery import shared_task
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
+from askpolis.db import get_db
 
 from askpolis.data_fetcher import FetchedDataRepository
 from askpolis.data_fetcher.abgeordnetenwatch import AbgeordnetenwatchDataFetcher
 
-engine = create_engine(os.getenv("DATABASE_URL") or "postgresql+psycopg://postgres@postgres:5432/askpolis-db")
-DbSession = sessionmaker(bind=engine)
 
 
 @shared_task(name="fetch_bundestag_from_abgeordnetenwatch")
 def fetch_bundestag_from_abgeordnetenwatch() -> None:
     bundestag_id = 5
-    session = DbSession()
+    session = next(get_db())
     try:
         data_fetcher = AbgeordnetenwatchDataFetcher(FetchedDataRepository(session))
         data_fetcher.fetch_election_programs(bundestag_id)
@@ -24,7 +19,7 @@ def fetch_bundestag_from_abgeordnetenwatch() -> None:
 
 @shared_task(name="cleanup_outdated_data")
 def cleanup_outdated_data() -> None:
-    session = DbSession()
+    session = next(get_db())
     try:
         FetchedDataRepository(session).delete_outdated_data()
     finally:
