+++ backend/src/askpolis/db/__init__.py
@@ -1,0 +1,3 @@
+from .dependencies import get_db
+
+__all__ = ["get_db"]
